chore: move users into VueX module
